refactor(imports): point all repair-request-create-intent importers to repair-request-deep-link
Eight production/test files plus the adoption test now import from the
renamed module. No behaviour change.

Importers updated (sed-replace of import path string only):
- src/components/mobile-equipment-list-item.tsx
- src/components/equipment/equipment-actions-menu.tsx
- src/components/assistant/AssistantPanel.tsx
- src/app/(app)/dashboard/page.tsx
- src/app/(app)/qr-scanner/page.tsx
- src/app/(app)/dashboard/__tests__/dashboard-no-legacy-dialog.test.tsx (vi.mock string)
- src/app/(app)/qr-scanner/__tests__/qr-scanner-no-legacy-dialog.test.tsx (vi.mock string)
- src/app/(app)/repair-requests/_hooks/useRepairRequestsDeepLink.ts

Adoption test renamed via git mv:
- src/lib/__tests__/repair-request-create-intent.adoption.test.ts
  → src/lib/__tests__/repair-request-deep-link.adoption.test.ts

Verify gates: verify:no-explicit-any green, typecheck green (0 errors),
focused test:run green (src/lib/__tests__/ 177 pass / 4 skipped).

Refs #338, #207.
