+++ src/app/(app)/dashboard/__tests__/dashboard-no-legacy-dialog.test.tsx
@@ -118,7 +118,7 @@ vi.mock("@/components/dashboard/RecentActivitiesCard", () => ({
     RecentActivitiesCard: () => <div data-testid="recent-activities-card" />,
 }))
 
-vi.mock("@/lib/repair-request-create-intent", () => ({
+vi.mock("@/lib/repair-request-deep-link", () => ({
     buildRepairRequestCreateIntentHref: (id?: number) =>
         `/repair-requests?equipmentId=${id}`,
 }))
