+++ src/app/(app)/dashboard/page.tsx
@@ -21,7 +21,7 @@ import { KPICards } from "@/components/dashboard/kpi-cards"
 import { DashboardTabs } from "@/components/dashboard/dashboard-tabs"
 import { QRScannerErrorBoundary } from "@/components/qr-scanner-error-boundary"
 import { useToast } from "@/hooks/use-toast"
-import { buildRepairRequestCreateIntentHref } from "@/lib/repair-request-create-intent"
+import { buildRepairRequestCreateIntentHref } from "@/lib/repair-request-deep-link"
 import { cn } from "@/lib/utils"
 import { isRegionalLeaderRole } from "@/lib/rbac"
 import type { Equipment } from "@/lib/data"
